Fixing use of std:max with float

+++ src/view_floating.cpp
@@ -981,7 +981,7 @@ int View::AdjustSlurCurve(Slur *slur, ArrayOfLayerElementPointPairs *spanningPoi
     int maxHeight = 0;
 
     // 0.2 for avoiding / by 0 (below)
-    float maxHeightFactor = std::max(0.2f, fabs(angle));
+    float maxHeightFactor = std::max(0.2f, fabsf(angle));
     maxHeight = dist / (maxHeightFactor * (TEMP_SLUR_CURVE_FACTOR
                                               + 5)); // 5 is the minimum - can be increased for limiting curvature
     if (posRatio) {
